add test with plate
